backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
curl_cffi==0.13.0
dnspython==2.7.0
fastapi==0.116.1
frozendict==2.4.6
h11==0.16.0
idna==3.10
motor==3.7.1
multitasking==0.0.12
numpy==2.3.2
pandas==2.3.1
peewee==3.18.2
platformdirs==4.3.8
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.47.2
ta==0.11.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
yfinance==0.2.65
